requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-ckeditor>=6.7,<7.0
dj-database-url>=2.2,<3.0
gunicorn>=22.0,<24.0
whitenoise>=6.8,<7.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<12.0
